Chore: Rename modules to utils to not mislead with node_modules
